# spring-boot2/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  description: |
    Apollo配置中心OpenAPI接口文档

    ## 认证方式

    所有API接口都需要通过Authorization header进行身份验证。

    ### 获取Token的方式：

    1. **Portal管理界面获取**：
       - 登录Apollo Portal管理界面
       - 进入"管理员工具" -> "开放平台授权管理"
       - 创建第三方应用，获取Token

    2. **Token格式**：
       ```
       Authorization: token_value
       ```

    3. **Token权限**：
       - Token具有对应应用的读写权限
       - 不同Token可能有不同的环境和命名空间权限
       - 建议为不同用途创建不同的Token

    ---

    ### 使用示例：
    ```bash
    curl -X GET "http://localhost:8070/openapi/v1/apps" \
         -H "Authorization: your_token_here"
    ```
  title: Apollo OpenAPI
  version: 1.0.0
servers:
- url: /
security:
- ApiKeyAuth: []
...
paths:
...
  /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/releases:
    post:
      deprecated: false
      description: 创建并发布当前命名空间的配置
      parameters:
      - description: 应用ID
        explode: false
        in: path
        name: appId
        required: true
        schema:
          type: string
        style: simple
      - description: 环境标识
        explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - description: 集群名称
        explode: false
        in: path
        name: clusterName
        required: true
        schema:
          type: string
        style: simple
      - description: 命名空间名称
        explode: false
        in: path
        name: namespaceName
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              allOf:
              - $ref: "#/components/schemas/NamespaceReleaseDTO"
              description: 创建发布请求体
              required:
              - releaseTitle
              - releasedBy
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              example:
                dataChangeCreatedBy: apollo
                dataChangeLastModifiedBy: apollo
                dataChangeCreatedTime: 2024-01-15T10:30:00.000Z
                dataChangeLastModifiedTime: 2024-01-15T10:30:00.000Z
                id: 12345
                appId: sample-app
                clusterName: default
                namespaceName: application
                name: Release v1.0.0
                configurations:
                  server.port: "8080"
                  spring.datasource.url: jdbc:mysql://localhost:3306/apollo
                  logging.level.root: INFO
                comment: 首次发布，包含基础配置
              schema:
                $ref: "#/components/schemas/OpenReleaseDTO"
          description: 发布创建成功
        "400":
          content:
            application/json:
              example:
                message: Params(releaseTitle and releasedBy) can not be empty
              schema:
                $ref: "#/components/schemas/_openapi_v1_apps_get_401_response"
          description: 发布参数错误
        "403":
          content:
            application/json:
              example:
                message: Access is denied
              schema:
                $ref: "#/components/schemas/_openapi_v1_apps_get_401_response"
          description: 权限不足
      summary: 创建发布
      tags:
      - Release Management
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Release Management
  /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/releases/latest:
    get:
      deprecated: false
      description: 查询命名空间最新活跃发布
      parameters:
      - description: 应用ID
        explode: false
        in: path
        name: appId
        required: true
        schema:
          type: string
        style: simple
      - description: 环境标识
        explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - description: 集群名称
        explode: false
        in: path
        name: clusterName
        required: true
        schema:
          type: string
        style: simple
      - description: 命名空间名称
        explode: false
        in: path
        name: namespaceName
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                dataChangeCreatedBy: apollo
                dataChangeLastModifiedBy: apollo
                dataChangeCreatedTime: 2024-01-15T10:30:00.000Z
                dataChangeLastModifiedTime: 2024-01-15T10:30:00.000Z
                id: 12345
                appId: sample-app
                clusterName: default
                namespaceName: application
                name: Release v1.2.0
                configurations:
                  server.port: "8080"
                  spring.datasource.url: jdbc:mysql://localhost:3306/apollo
                  logging.level.root: INFO
                  app.version: 1.2.0
                comment: 最新生产发布版本，包含性能优化
              schema:
                $ref: "#/components/schemas/OpenReleaseDTO"
          description: 成功获取最新活跃发布
        "404":
          content:
            application/json:
              example:
                message: No active release found for this namespace
              schema:
                $ref: "#/components/schemas/_openapi_v1_apps_get_401_response"
          description: 未找到活跃发布
      summary: 获取最新活跃发布
      tags:
      - Release Management
      x-accepts:
      - application/json
      x-tags:
      - tag: Release Management
...
  /openapi/v1/envs/{env}/releases/{releaseId}/rollback:
    put:
      deprecated: false
      description: 回滚到指定的发布版本
      parameters:
      - description: 环境标识
        explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - description: 发布ID
        explode: false
        in: path
        name: releaseId
        required: true
        schema:
          type: integer
        style: simple
      - description: 操作人用户名
        explode: true
        in: query
        name: operator
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                properties: {}
                type: object
          description: 发布回滚成功
      summary: 回滚发布
      tags:
      - Release Management
      x-accepts:
      - application/json
      x-tags:
      - tag: Release Management
  /openapi/v1/envs/{env}/releases/{releaseId}:
    get:
      deprecated: false
      description: "GET /openapi/v1/envs/{env}/releases/{releaseId}"
      parameters:
      - description: 环境标识
        explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - description: 发布ID
        explode: false
        in: path
        name: releaseId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              example:
                dataChangeCreatedBy: ""
                dataChangeLastModifiedBy: ""
                dataChangeCreatedTime: ""
                dataChangeLastModifiedTime: ""
                id: 0
                appId: ""
                clusterName: ""
                namespaceName: ""
                name: ""
                configurations:
                  ? ""
                  : ""
                comment: ""
              schema:
                $ref: "#/components/schemas/OpenReleaseDTO"
          description: 成功获取发布详情
      summary: 获取发布详情
      tags:
      - Release Management
      x-accepts:
      - application/json
      x-tags:
      - tag: Release Management
  /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/releases/all:
    get:
      deprecated: false
      description: "GET /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/releases/all"
      parameters:
      - description: 应用ID
        explode: false
        in: path
        name: appId
        required: true
        schema:
          type: string
        style: simple
      - description: 环境标识
        explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - description: 集群名称
        explode: false
        in: path
        name: clusterName
        required: true
        schema:
          type: string
        style: simple
      - description: 命名空间名称
        explode: false
        in: path
        name: namespaceName
        required: true
        schema:
          type: string
        style: simple
      - description: 页码，从0开始
        example: 0
        explode: true
        in: query
        name: page
        required: true
        schema:
          type: integer
        style: form
      - description: 每页数量
        example: 5
        explode: true
        in: query
        name: size
        required: true
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
              - baseInfo:
                  dataChangeCreatedBy: ""
                  dataChangeLastModifiedBy: ""
                  dataChangeCreatedByDisplayName: ""
                  dataChangeLastModifiedByDisplayName: ""
                  dataChangeCreatedTime: ""
                  dataChangeLastModifiedTime: ""
                  id: 0
                  releaseKey: ""
                  name: ""
                  appId: ""
                  clusterName: ""
                  namespaceName: ""
                  configurations: ""
                  comment: ""
                  isAbandoned: false
                items:
                - key: ""
                  value: ""
              schema:
                items:
                  $ref: "#/components/schemas/ListReleaseBO"
                type: array
          description: 成功获取发布列表
      summary: 获取所有发布（分页）
      tags:
      - Release Management
      x-accepts:
      - application/json
      x-tags:
      - tag: Release Management
  /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/releases/active:
    get:
      deprecated: false
      description: "GET /openapi/v1/envs/{env}/apps/{appId}/clusters/{clusterName}/namespaces/{namespaceName}/releases/active"
      parameters:
      - description: 应用ID
        explode: false
        in: path
        name: appId
        required: true
        schema:
          type: string
        style: simple
      - description: 环境标识
        explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - description: 集群名称
        explode: false
        in: path
        name: clusterName
        required: true
        schema:
          type: string
        style: simple
      - description: 命名空间名称
        explode: false
        in: path
        name: namespaceName
        required: true
        schema:
          type: string
        style: simple
      - description: 页码，从0开始
        example: 0
        explode: true
        in: query
        name: page
        required: true
        schema:
          type: integer
        style: form
      - description: 每页数量
        example: 5
        explode: true
        in: query
        name: size
        required: true
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
              - dataChangeCreatedBy: ""
                dataChangeLastModifiedBy: ""
                dataChangeCreatedTime: ""
                dataChangeLastModifiedTime: ""
                id: 0
                appId: ""
                clusterName: ""
                namespaceName: ""
                name: ""
                configurations:
                  ? ""
                  : ""
                comment: ""
              schema:
                items:
                  $ref: "#/components/schemas/OpenReleaseDTO"
                type: array
          description: 成功获取活跃发布列表
      summary: 获取活跃发布（分页）
      tags:
      - Release Management
      x-accepts:
      - application/json
      x-tags:
      - tag: Release Management
  /openapi/v1/envs/{env}/releases/compare:
    get:
      deprecated: false
      description: "GET /openapi/v1/envs/{env}/releases/compare"
      parameters:
      - description: 环境标识
        explode: false
        in: path
        name: env
        required: true
        schema:
          type: string
        style: simple
      - description: 基准发布ID
        explode: true
        in: query
        name: baseReleaseId
        required: true
        schema:
          type: integer
        style: form
      - description: 对比发布ID
        explode: true
        in: query
        name: toCompareReleaseId
        required: true
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              example:
                changes:
                - type: ""
                  entity:
                    firstEntity:
                      key: ""
                      value: ""
                    secondEntity:
                      key: ""
                      value: ""
              schema:
                $ref: "#/components/schemas/_openapi_v1_envs__env__releases_compare_get_200_response"
          description: 成功返回发布差异
      summary: 对比发布
      tags:
      - Release Management
      x-accepts:
      - application/json
      x-tags:
      - tag: Release Management
...
components:
  schemas:
...
    MapString:
      description: 字符串映射对象，用于表示键值对的配置数据结构
      example:
        key: key
      properties:
        key:
          type: string
      type: object
    OpenReleaseDTO:
      description: Apollo发布信息数据传输对象，表示一次配置发布的完整信息
      example:
        dataChangeCreatedTime: dataChangeCreatedTime
        dataChangeLastModifiedBy: dataChangeLastModifiedBy
        configurations:
          key: key
        appId: appId
        clusterName: clusterName
        dataChangeCreatedBy: dataChangeCreatedBy
        name: name
        comment: comment
        id: 0
        dataChangeLastModifiedTime: dataChangeLastModifiedTime
        namespaceName: namespaceName
      properties:
        dataChangeCreatedBy:
          description: 发布创建者用户名，记录是谁创建了这次发布
          type: string
        dataChangeLastModifiedBy:
          description: 发布最后修改者用户名，记录最后一次修改发布信息的用户
          type: string
        dataChangeCreatedTime:
          description: 发布创建时间，ISO 8601格式的时间戳
          type: string
        dataChangeLastModifiedTime:
          description: 发布最后修改时间，ISO 8601格式的时间戳
          type: string
        id:
          description: 发布记录的唯一标识符，系统自动生成
          format: int64
          type: integer
        appId:
          description: 所属应用的唯一标识符
          type: string
        clusterName:
          description: 所属集群的名称
          type: string
        namespaceName:
          description: 所属命名空间的名称
          type: string
        name:
          description: 发布名称，用于标识这次发布的版本或描述
          type: string
        configurations:
          $ref: "#/components/schemas/MapString"
        comment:
          description: 发布备注，描述本次发布的变更内容和目的
          type: string
      type: object
    NamespaceReleaseDTO:
      description: 命名空间发布请求数据传输对象，用于创建新的配置发布
      example:
        releaseComment: releaseComment
        releasedBy: releasedBy
        releaseTitle: releaseTitle
        isEmergencyPublish: true
      properties:
        releaseTitle:
          description: 发布标题，用于标识这次发布的名称或版本号
          type: string
        releaseComment:
          description: 发布备注，详细描述本次发布的变更内容和目的
          type: string
        releasedBy:
          description: 发布操作者用户名，记录是谁执行了这次发布操作
          type: string
        isEmergencyPublish:
          description: 是否为紧急发布，紧急发布可能会跳过某些审核流程
          type: boolean
      type: object
...
    _openapi_v1_apps_get_401_response:
      example:
        message: message
      properties:
        message:
          type: string
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: |
        Apollo OpenAPI访问令牌
        **获取方式：**
        1. 登录Apollo Portal管理界面
        2. 进入"管理员工具" -> "开放平台授权管理"
        3. 创建第三方应用，获取Token
        **使用格式：**
        ```
        Authorization: your_token_here
        ```
        **权限说明：**
        - Token具有对应应用的读写权限
        - 不同Token可能有不同的环境和命名空间权限
        - 建议为不同用途创建不同的Token
      in: header
      name: Authorization
      type: apiKey